web/app/py-requirements.txt
Django==2.1.5
djangorestframework==3.9.1
Markdown==3.0.1
pytz==2018.9

